Calculate total weight of truffles grouped by specified columns

This function calculates the total weight of truffles grouped by specified columns in a given data frame.

Usage

weight_truffles_by(dbtruffle, ...)

Arguments

dbtruffle

A data frame containing information about truffles.

...

Columns to group by, passed as arguments.

Value

A summarized data frame containing the total weight of truffles grouped by the specified columns.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

conn <- DBI::dbConnect(
   RSQLite::SQLite(),
   system.file(dbname = "chenes_truffe.sqlite", package = "truffles")
 )
truffes <- DBI::dbReadTable(conn, name = "truffe")

weight_truffles_by(truffes, annee = lubridate::year(as.Date(date_found)))
#> # A tibble: 3 × 2
#>   annee weight
#>   <dbl>  <int>
#> 1  2022   5583
#> 2  2023   3837
#> 3  2024    493

truffes_chene <- truffes |>
  inner_join(DBI::dbReadTable(conn, name = "chenes"), by = "idoak")
#> Warning: Detected an unexpected many-to-many relationship between `x` and `y`.
#> ℹ Row 4 of `x` matches multiple rows in `y`.
#> ℹ Row 1 of `y` matches multiple rows in `x`.
#> ℹ If a many-to-many relationship is expected, set `relationship =
#>   "many-to-many"` to silence this warning.
weight_truffles_by(truffes_chene, annee = lubridate::year(as.Date(date_found)), type)
#> # A tibble: 5 × 3
#>   annee type   weight
#>   <dbl> <chr>   <int>
#> 1  2022 Green     651
#> 2  2022 Normal   6162
#> 3  2023 Green     407
#> 4  2023 Normal   3770
#> 5  2024 Normal    493